Reorganize examples, tests and Dockerfiles
