+++ .github/workflows/test.yaml
@@ -30,23 +30,6 @@ jobs:
     strategy:
       matrix:
         include:
-          - platform: linux/arm64
-            dockerfile: Dockerfile.arm
-            tag: nntbn:arm64
-            test: arm
-
-          # No NEON support for armv7
-          # - platform: linux/arm/v7
-          #   dockerfile: Dockerfile.arm
-          #   tag: nntbn:armv7
-          #   test: arm
-
-          # No NEON support for armv8
-          # - platform: linux/arm/v8
-          #   dockerfile: Dockerfile.arm
-          #   tag: nntbn:armv8
-          #   test: arm
-
           - platform: linux/amd64
             dockerfile: Dockerfile.generic
             tag: nntbn:amd64
@@ -61,6 +44,13 @@ jobs:
             dockerfile: Dockerfile.generic
             tag: nntbn:generic-armv8
             test: generic
+
+          # TODO: Arm Neon is not available on ubuntu:latest image, try another one
+          - platform: linux/arm64
+            dockerfile: Dockerfile.neon
+            tag: nntbn:arm64
+            test: arm-neon
+
     steps:
       - name: Docker Setup QEMU
         uses: docker/setup-qemu-action@v3.0.0
@@ -84,9 +74,10 @@ jobs:
       - name: Build and Run Test
         run: |
           docker buildx build --platform ${{ matrix.platform }} -f scripts/docker/${{ matrix.dockerfile }} -t ${{ matrix.tag }} . --load
-          if [ "${{ matrix.test }}" == "arm" ]; then
-            docker run --platform ${{ matrix.platform }} ${{ matrix.tag }} /nn/build/tests/arch/arm/cmsis/neuron
+          if [ "${{ matrix.test }}" == "arm-neon" ]; then
             docker run --platform ${{ matrix.platform }} ${{ matrix.tag }} /nn/build/tests/arch/arm/neon/neuron
+            docker run --platform ${{ matrix.platform }} ${{ matrix.tag }} /nn/build/tests/arch/arm/neon/dot_product
           else
             docker run --platform ${{ matrix.platform }} ${{ matrix.tag }} /nn/build/tests/arch/generic/neuron
+            docker run --platform ${{ matrix.platform }} ${{ matrix.tag }} /nn/build/tests/arch/generic/dot_product
           fi
