--- scripts/docker/Dockerfile.arm
@@ -1,28 +1,0 @@
-# Example usage:
-# Build:
-#   docker buildx build --platform linux/arm64 -f scripts/docker/Dockerfile.arm -t nntbn:arm . --load
-# Run:
-#   docker run --rm -it --platform linux/arm64 nntbn:arm /nn/build/tests/arch/arm/cmsis/neuron
-#   docker run --rm -it --platform linux/arm64 nntbn:arm /nn/build/tests/arch/arm/neon/neuron
-
-# Build
-FROM ubuntu:latest as build
-ENV DEBIAN_FRONTEND=noninteractive
-RUN apt-get update && apt-get install -y \
-    build-essential \
-    clang \
-    lld \
-    gcc-arm-none-eabi \
-    libnewlib-arm-none-eabi
-COPY . /nn
-WORKDIR /nn
-RUN make build-examples ARCH=arm LDFLAGS="-lm"
-RUN make test ARCH=arm LDFLAGS="-lm"
-
-# Runtime
-FROM ubuntu:latest
-WORKDIR /nn
-COPY --from=build /nn /nn
-
-ENTRYPOINT ["/bin/bash", "-c"]
-CMD ["/bin/bash"]
