+++ scripts/docker/Dockerfile.arm-cmsis-dsp
@@ -1,0 +1,28 @@
+# Example usage:
+# Build:
+#   docker buildx build --platform linux/arm64 -f scripts/docker/Dockerfile.arm-cmsis-dsp -t nntbn:arm-cmsis-dsp . --load
+# Run:
+#   docker run --rm -it --platform linux/arm64 nntbn:arm-cmsis-dsp /nn/build/tests/arch/arm/cmsis-dsp/neuron
+#   docker run --rm -it --platform linux/arm64 nntbn:arm-cmsis-dsp /nn/build/examples/arch/arm/cmsis-dsp/dot_product
+
+# Build
+FROM ubuntu:latest as build
+ENV DEBIAN_FRONTEND=noninteractive
+RUN apt-get update && apt-get install -y \
+    build-essential \
+    clang \
+    lld \
+    gcc-arm-none-eabi \
+    libnewlib-arm-none-eabi
+COPY . /nn
+WORKDIR /nn
+RUN make build-examples ARCH=arm TECH=cmsis-dsp LDFLAGS="-lm"
+RUN make test ARCH=arm TECH=cmsis-dsp LDFLAGS="-lm"
+
+# Runtime
+FROM ubuntu:latest
+WORKDIR /nn
+COPY --from=build /nn /nn
+
+ENTRYPOINT ["/bin/bash", "-c"]
+CMD ["/bin/bash"]
